Integrate router components
