+++ src/App.css
@@ -1,5 +1,3 @@
-/* App.css */
-
 /* Dark Mode color palette */
 :root {
   --bg-color: #1a1a1a; /* Dark background color for the main content */
@@ -13,37 +11,4 @@ body {
   font-family: Arial, sans-serif;
   background-color: var(--bg-color); /* Dark background color for dark mode */
   color: var(--text-color); /* Light text color for dark mode */
-}
-
-.app-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px;
-  background-color: var(--bg-color); /* Dark background color for the header */
-  color: var(--text-color); /* Light text color for the header */
-}
-
-nav button {
-  padding: 8px 16px;
-  border: none;
-  background-color: var(--bg-color); /* Dark background color for buttons in the header */
-  color: var(--text-color); /* Light text color for buttons in the header */
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-nav button:hover {
-  background-color: var(--primary-color); /* Blue background color on hover */
-}
-
-.main-content {
-  padding: 16px;
-}
-
-.app-footer {
-  text-align: center;
-  padding: 16px;
-  background-color: var(--bg-color); /* Dark background color for the footer */
-  color: var(--text-color); /* Light text color for the footer */
 }
